Add support for replacement of ampersands from company names
WHMCS should be HTML encoding the values they send to Plesk via API,
but they don’t. Since the hooks don’t allow us to filter the data prior
to sending to Plesk, we can’t simply run htmlentities() on it as it
must be saved to the client profile. Instead find and replace both
encoded and non-encoded ‘&’ values with the word ’and’ then send to
Plesk.