Plesk is picky about which symbols it counts as secure
One of the recently generated passwords managed to only have
parentheses “)” and “(“ which Plesk deemed insecure (silly). This
change ensures the symbols allowed match with the set that Plesk likes.

I should note that it’s definitely possible that a password could be
generated without symbols (although not probable) and thusly still
cause a problem. Fixing this would require some modifications to the
randomPassword function.